+++ library/src/main/java/L/AndroidLogTool.java
@@ -14,11 +14,9 @@
  * limitations under the License.
  */
 
-package utils;
+package L;
 
 import android.util.Log;
-
-import interfa.LogTool;
 
 public class AndroidLogTool implements LogTool {
   @Override public void d(String tag, String message) {