+++ library/src/main/java/L/ArrayUtil.java
@@ -14,11 +14,13 @@
  * limitations under the License.
  */
 
-package utils;
+package L;
 
 import android.util.Pair;
 
 import java.util.Arrays;
+
+import L.SystemUtil;
 
 /**
  * Created by pengwei08 on 2015/7/25.