+++ library/src/main/java/L/L.java
@@ -14,9 +14,7 @@
  * limitations under the License.
  */
 
-package utils;
-
-import interfa.Printer;
+package L;
 
 /**
  * @author L.M