+++ library/src/main/java/L/LogLevel.java
@@ -14,37 +14,17 @@
  * limitations under the License.
  */
 
-package interfa;
+package L;
 
-import utils.Settings;
+public enum LogLevel {
 
-public interface Printer {
+  /**
+   * Prints all logs
+   */
+  FULL,
 
-  Printer t(String tag, int methodCount);
-
-  Settings init(String tag);
-
-  Settings getSettings();
-
-  void d(String message, Object... args);
-
-  void e(String message, Object... args);
-
-  void e(Throwable throwable, String message, Object... args);
-
-  void w(String message, Object... args);
-
-  void i(String message, Object... args);
-
-  void v(String message, Object... args);
-
-  void wtf(String message, Object... args);
-
-  void json(String json);
-
-  void xml(String xml);
-
-  void object(Object object);
-
-  void clear();
+  /**
+   * No log will be printed
+   */
+  NONE
 }