+++ library/src/main/java/commonadapter/CommonAdapter.java
@@ -1,0 +1,77 @@
+/*
+ * Copyright 2015 LuMeng
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package commonadapter;
+
+import android.content.Context;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.BaseAdapter;
+
+import java.util.List;
+
+import commonadapter.ViewHolder;
+
+/**
+ * We just leave a {@link #convert(ViewHolder, Object)} method for inheritor.
+ * All of the things you should do is ~ setText or setImage
+ *
+ * @author L.M
+ * @version 1.0.2
+ */
+public abstract class CommonAdapter<T> extends BaseAdapter {
+
+    protected LayoutInflater layoutInflater;
+    protected Context context;
+    protected List<T> datas;
+    protected final int itemLayoutId;
+
+    public CommonAdapter(Context context, List<T> datas, int itemLayoutId) {
+        this.context = context;
+        this.layoutInflater = LayoutInflater.from(context);
+        this.datas = datas;
+        this.itemLayoutId = itemLayoutId;
+    }
+
+    @Override
+    public int getCount() {
+        return datas.size();
+    }
+
+    @Override
+    public T getItem(int position) {
+        return datas.get(position);
+    }
+
+    @Override
+    public long getItemId(int position) {
+        return position;
+    }
+
+    @Override
+    public View getView(int position, View convertView, ViewGroup parent) {
+        final ViewHolder viewHolder = getViewHolder(position, convertView, parent);
+        convert(viewHolder, getItem(position));
+        return viewHolder.getConvertView();
+    }
+
+    public abstract void convert(ViewHolder helper, T item);
+
+    private ViewHolder getViewHolder(int position, View convertView, ViewGroup parent) {
+        return ViewHolder.get(context, convertView, parent, itemLayoutId, position);
+    }
+}