+++ CHANGELOG.md
@@ -8,3 +8,7 @@
 - Fri Oct 23
 
     add **CircleImageView**, **BitmapUtils**, **ViewHolder**
+    
+- Sat Oct 24
+
+    add **CommonAdapter** cooperation with **ViewHolder** below